+++ tox.ini
@@ -163,7 +163,7 @@ commands =
     python --version
     pip install -U --target {envsitepackagesdir}/azure-cli-extensions/azure-iot .
     az -v
-    pytest -vv ./azext_iot/tests/adr/test_adr_certmgmt_lifecycle_int.py \
+    pytest -vv ./azext_iot/tests/adr/test_adr_device_int.py \
         -k test_adr_device_certmgmt_lifecycle \
         --cov=azext_iot --cov-append --cov-config .coveragerc --junitxml=junit/test-iotext-int.xml {posargs}
